Significant code clean-up in preparation for the release of 0.3.10.93

+++ AW.Core/AW.DataStructures.cs
@@ -77,8 +77,8 @@ namespace AW
 
         public short Speed
         {
-            get { return (short)speed; }
-            set { speed = (short)value; }
+            get { return speed; }
+            set { speed = value; }
         }
 
         public byte Pause
